requirements.txt
discord==1.0.1
discord.py==1.7.2
psutil==5.8.0
PyNaCl==1.4.0
python-dotenv==0.18.0
python-dotenv-run==0.1.4
requests==2.25.1